move around samples
