+++ openai_agents/basic/run_worker.py
@@ -1,0 +1,43 @@
+from __future__ import annotations
+
+import asyncio
+from datetime import timedelta
+
+from temporalio.client import Client
+from temporalio.contrib.openai_agents import ModelActivityParameters, OpenAIAgentsPlugin
+from temporalio.worker import Worker
+
+from openai_agents.basic.activities.get_weather_activity import get_weather
+from openai_agents.basic.workflows.hello_world_workflow import HelloWorldAgent
+from openai_agents.basic.workflows.tools_workflow import ToolsWorkflow
+
+
+async def main():
+    # Create client connected to server at the given address
+    client = await Client.connect(
+        "localhost:7233",
+        plugins=[
+            OpenAIAgentsPlugin(
+                model_params=ModelActivityParameters(
+                    start_to_close_timeout=timedelta(seconds=30)
+                )
+            ),
+        ],
+    )
+
+    worker = Worker(
+        client,
+        task_queue="openai-agents-task-queue",
+        workflows=[
+            HelloWorldAgent,
+            ToolsWorkflow,
+        ],
+        activities=[
+            get_weather,
+        ],
+    )
+    await worker.run()
+
+
+if __name__ == "__main__":
+    asyncio.run(main())
