+++ openai_agents/research_bot/workflows/research_bot_workflow.py
@@ -1,6 +1,6 @@
 from temporalio import workflow
 
-from openai_agents.workflows.research_agents.research_manager import ResearchManager
+from openai_agents.research_bot.agents.research_manager import ResearchManager
 
 
 @workflow.defn
